requirements.txt
streamlit
boto3==1.34.91
python-docx==1.1.0
pdfplumber==0.11.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
PyPDF2==3.0.1
python-dotenv==1.0.1